requirements.txt
streamlit
sqlalchemy
pydantic
yahoo_fin
python-dotenv
passlib
streamlit_authenticator
pandas
numpy
